+++ torsiondrive/tests/test_extra_constraints.py
@@ -210,17 +210,11 @@ def test_build_terachem_constraint_string():
     }
     constraints_string = build_terachem_constraint_string(constraints_dict)
     # validate the constraints string
-    assert constraints_string.strip() == '\n'.join(['$constraint_freeze', 'xyz 1-3,6-7', 'bond 1_5', '$end\n', '$constraint_set', 'bond 1.0 1_4',
-        'angle 30.0 2_1_4', 'dihedral 60.0 1_2_3_4', '$end'])
+    assert constraints_string.strip() == '\n'.join(['$constraint_freeze', 'xyz 1-3,6-7', 'bond 1_5', '$end\n',
+        '$constraint_set', 'bond 1.0 1_4', 'angle 30.0 2_1_4', 'dihedral 60.0 1_2_3_4', '$end'])
     # test with dihedral_idx_values
     dihedral_idx_values=[(1,2,3,4,90.0), (2,3,4,5,120.0)]
     constraints_string2 = build_terachem_constraint_string(constraints_dict, dihedral_idx_values=dihedral_idx_values)
-    assert constraints_string2.strip() == '\n'.join(['$constraint_freeze', 'xyz 1-3,6-7', 'bond 1_5', '$end\n', '$constraint_set', 'bond 1.0 1_4',
-        'angle 30.0 2_1_4', 'dihedral 60.0 1_2_3_4', 'dihedral 90.0 2_3_4_5', 'dihedral 120.0 3_4_5_6', '$end'])
-
-def test_reproduce_extra_constraints_example():
-    """
-    Test reproducing examples/extra_constraints
-    """
-    # to be implemented
-    pass
+    assert constraints_string2.strip() == '\n'.join(['$constraint_freeze', 'xyz 1-3,6-7', 'bond 1_5', '$end\n',
+        '$constraint_set', 'bond 1.0 1_4', 'angle 30.0 2_1_4', 'dihedral 60.0 1_2_3_4',
+        'dihedral 90.0 2_3_4_5', 'dihedral 120.0 3_4_5_6', '$end'])
